Backend/requirements.txt
fastapi
uvicorn
langchain
sqlalchemy
psycopg2-binary
python-dotenv
pandas
openpyxl
qdrant-client
sentence-transformers
langchain-community
langchain-huggingface
langchain-qdrant
langchain-google-genai
langchain-groq
